+++ packages/rnv-engine-rn/src/sdks/index.js
@@ -1,7 +1,5 @@
 import * as SDKXcode from './sdk-xcode';
-import * as SDKAndroid from './sdk-android';
 
 export {
-    SDKAndroid,
     SDKXcode
 };
