+++ packages/rnv-engine-rn/src/tasks/task.rnv.package.js
@@ -1,5 +1,6 @@
 import { TaskManager, Constants, Logger, PlatformManager, Common } from 'rnv';
-import { SDKAndroid, SDKXcode } from '../sdks';
+import { packageAndroid } from '@rnv/sdk-android';
+import { SDKXcode } from '../sdks';
 
 const { logErrorPlatform } = PlatformManager;
 const { logTask } = Logger;
@@ -16,7 +17,7 @@ const {
 } = Constants;
 const { getConfigProp } = Common;
 const { packageBundleForXcode } = SDKXcode;
-const { packageAndroid } = SDKAndroid;
+
 const { executeOrSkipTask, shouldSkipTask } = TaskManager;
 
 export const taskRnvPackage = async (c, parentTask, originTask) => {
