+++ packages/rnv-engine-rn/src/tasks/task.rnv.run.js
@@ -1,6 +1,7 @@
 import { TaskManager, Constants, Logger, PlatformManager, Common } from 'rnv';
+import { packageAndroid, runAndroid } from '@rnv/sdk-android';
 import { startBundlerIfRequired, waitForBundlerIfRequired } from '../commonEngine';
-import { SDKAndroid, SDKXcode } from '../sdks';
+import { SDKXcode } from '../sdks';
 
 const { getConfigProp } = Common;
 const { logErrorPlatform } = PlatformManager;
@@ -17,7 +18,6 @@ const {
     PARAMS
 } = Constants;
 const { runXcodeProject } = SDKXcode;
-const { packageAndroid, runAndroid } = SDKAndroid;
 const { executeOrSkipTask, shouldSkipTask } = TaskManager;
 
 export const taskRnvRun = async (c, parentTask, originTask) => {
