+++ packages/rnv-sdk-android/src/supportFiles/AndroidManifestFULL.json
