+++ packages/rnv-sdk-android/src/supportFiles/AndroidManifest_android.json
