+++ packages/rnv-sdk-android/src/supportFiles/AndroidManifest_androidtv.json
