+++ packages/rnv-sdk-android/src/supportFiles/AndroidManifest_androidwear.json
