+++ packages/rnv-sdk-android/src/supportFiles/AndroidManifest_firetv.json
