+++ packages/rnv-sdk-android/src/supportFiles/app.build.gradle.json
