--- packages/rnv-sdk-android/src/types.d.ts
@@ -1,1 +1,0 @@
-declare module 'rnv'
