+++ packages/rnv-sdk-android/src/xmlValuesParser.js
