+++ packages/rnv-sdk-android/types.d.ts
@@ -1,0 +1,3 @@
+declare module 'rnv'
+declare module 'inquirer'
+declare module 'shelljs'
